split file impl for database
